+++ spec/fixtures/packages/some-package/index-hardcoded-name.js
@@ -1,0 +1,7 @@
+'use babel'
+
+module.exports = {
+  activate() {
+    return require('../../../../').install('some-package')
+  }
+}